Put the user defined symbol filter in last place
Because checking the symbol against its reflection, or even its
existence, symbols that are broken will crash the application. This can
happen when a class that is inspected is extending a class that does not
exist, or in similar conditions.

Putting the user defined symbol filter last ensures the package can
configure symbol exclusion rules that are applied before the broken
symbol is loaded in. A broken symbol can be a feature for some packages,
so it should not be something that has to be fixed in order for the
dependency guard to successfully run. I.e.: A mocked class for the
Symfony Serializer to test what happens when you deserialize to a class
that is missing its parent.