Reduce responsibilities of regression test
The regression test is no longer concerned with running processes or
cleaning up filesystem entries recursively.

A reference to the issue is added to the docblock of the regression
test.

Assumptions about the presence of `php`, `composer` and `composer.phar`
have been solved by using the `PHP_BINARY` constant and by chaining
calls to the `which` UNIX command.