Fix #25 - Out of memory errors
When a ReflectionClass from BetterReflection was requested statically,
it created a new instance of BetterReflection under the hood. This
instance of BetterReflection created a new PHP parser under the hood,
every time. So, every time a PHP symbol was reflected upon, the whole
parser was instantiated, which was really memory consuming.

In the case of issue #25, this resulted in an out of memory error,
requesting over 1GB of memory to run Dependency Guard.

By preventing zealous object instantiation, the memory consumption is
now below a set threshold of 15MB of memory for the package that was
affected in issue #25.