Support older version of BetterReflection
Version 3 of BetterReflection requires PHP parser 4, which in turn is
also required by PHPStan. PHPStan has a dependency on Symfony Console,
which supports 3 | 4 in more recent versions, and goes back to 2 in
older versions, which in turn force PHP parser 3 and thus becomes
incompatible with the newer version of BetterReflection.

With the new changes, Magento 2 projects and modules can be statically
analyzed by DependencyGuard.