Bugfix: resolved the infinite loop issue
This commit contains a proposal for a fix, for the infinite loop issue #33.

Instead of using the built-in composer getDependents function with the recurse option on true,
a new function will be introduced. This function will prevent an inifinite loop from occuring.
It does so by storing already resolved dependency information in a local property.

The test-case has also been adjusted, because with this change, there are no errors returned from the test-case.

A dependency on the package symfony/options-resolver has also been added to the dependency guard package itself.
The package was missing and dependency-guard was warning about this missing dependency.