Deprecate filtering and resolving using dependents
When DependencyGuard is expecting resolved dependents, it is expecting
package links of the "require" type. Leveraging Composer to determine
the dependents will result in a list with all types of package link,
which causes DependencyGuard to make assumptions using a wrong
dependency graph.

See https://getcomposer.org/doc/04-schema.md#package-links for more
information on what package links are.